pyproject.toml
[tool.poetry]
name = "jobber"
version = "0.1.0"
description = ""
authors = ["nischalj10 <nischalj10@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
anthropic = "^0.40.0"
pydantic = "^2.8.2"
pytest-playwright = "^0.5.1"
pdfplumber = "0.11.2"
typing-extensions = "^4.12.2"
ruff = "^0.5.6"
playwright-stealth = "^1.0.6"
setuptools = "^72.1.0"
python-json-logger = "^2.0.7"
aiohttp = "^3.10.2"
colorama = "^0.4.6"
termcolor = "^2.4.0"
tabulate = "^0.9.0"
nltk = "^3.9.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
